sdk: Move `Presigner` to `signer` module
